feat: Renombrar pantalla de creación de itinerarios y reorganizar navegación; eliminar HomeScreen
